refactor Examples into directory-based modules with co-located styles
