+++ packages/oc-docs/src/components/Docs/Item/Examples/ResponseExamples/StyledWrapper.tsx
